Update context manager for Client
On one hand this commit adds a nested context counter, to allow entering
and exiting the context multiple times, closing the node only upon
exiting the last context. On the other hand, this commit adds an async
alias for the synchronous context manager, to support chaining async
context managers (`async with Client(...) as c, Something(...) as s):`.