♻️ Refactor DSDL handling to use sc-packaged-dsdl
The sc-packaged-dsdl package [1] was created to simplify distribution of
(commonly used) DSDL namespaces. This enables us to drop the hacky
compilation and injection of DSDL types into the Python path, in favor
of simply installing the namespace package from GitHub.

With this change, the `cyphal install` command has been made obsolete.

[1]: https://github.com/starcopter/packaged-dsdl-py